+++ src/distributions/covers/CoverStack.h
@@ -43,7 +43,7 @@ class CoverStack: public multiset<StackEntry<T>>
       typename const set<T>::const_iterator& iterIn,
       const CoverTableau& tableau);
 
-    unsigned prune(const CoverTableau& solution);
+    size_t prune(const CoverTableau& solution);
 
 };
