+++ src/distributions/covers/CoverStore.h
@@ -64,7 +64,7 @@ class CoverStore
 
     set<Cover>::const_iterator end() const;
 
-    unsigned size() const;
+    size_t size() const;
 
     string str() const;
 };
